Update to imagej-modelzoo 0.9.2

+++ src/main/java/de/csbdresden/n2v/command/UpgradeN2VModelCommand.java
@@ -28,11 +28,11 @@
  */
 package de.csbdresden.n2v.command;
 
-import de.csbdresden.n2v.train.N2VConfig;
 import de.csbdresden.n2v.train.N2VModelSpecification;
+import io.bioimage.specification.io.SpecificationWriter;
 import net.imagej.modelzoo.ModelZooArchive;
 import net.imagej.modelzoo.ModelZooService;
-import net.imagej.modelzoo.specification.ModelSpecification;
+import io.bioimage.specification.ModelSpecification;
 import org.scijava.Context;
 import org.scijava.ItemIO;
 import org.scijava.command.Command;
@@ -43,11 +43,15 @@ import org.scijava.plugin.Plugin;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.DirectoryStream;
 import java.nio.file.FileSystem;
 import java.nio.file.FileSystems;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.PathMatcher;
 import java.nio.file.StandardCopyOption;
+import java.util.Enumeration;
+import java.util.zip.ZipEntry;
 
 import static org.scijava.widget.FileWidget.DIRECTORY_STYLE;
 
@@ -83,7 +87,7 @@ public class UpgradeN2VModelCommand implements Command {
 	ModelZooArchive tryUpgrade(File modelFile, File destinationFolder, String destinationFileName) {
 		ModelZooArchive model = null;
 		try {
-			model = modelZooService.open(modelFile);
+			model = modelZooService.io().open(modelFile);
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
@@ -106,27 +110,23 @@ public class UpgradeN2VModelCommand implements Command {
 
 	private ModelZooArchive upgrade(ModelZooArchive model, File destinationFolder, String destinationFileName) throws IOException {
 		ModelSpecification oldSpec = model.getSpecification();
-		File destination = new File(destinationFolder, destinationFileName + ".bioimage.io.zip");
+		File destination = new File(destinationFolder, destinationFileName + ".model.bioimage.io.zip");
 		if (destination.getAbsolutePath().equals(getAbsolutePath(model.getLocation()))) {
 			logService.error("Destination file cannot be the same as the deprecated model ZIP file");
 			return model;
 		}
 		Files.copy(new File(model.getLocation().getURI()).toPath(), destination.toPath(), StandardCopyOption.REPLACE_EXISTING);
 		N2VModelSpecification newSpec = new N2VModelSpecification();
-		newSpec.setMeta();
-		newSpec.setInputsOutputs(new N2VConfig());
-		newSpec.setName(oldSpec.getName());
-		newSpec.setDescription(oldSpec.getDescription());
-		newSpec.setAuthors(oldSpec.getAuthors());
-		newSpec.setTags(oldSpec.getTags());
-		newSpec.getPredictionPreprocessing().addAll(oldSpec.getPredictionPreprocessing());
-		newSpec.getPredictionPostprocessing().addAll(oldSpec.getPredictionPostprocessing());
+		newSpec.read(oldSpec);
 		try (FileSystem fileSystem = FileSystems.newFileSystem(destination.toPath(), null)) {
-			Path specPath = fileSystem.getPath(oldSpec.getModelFileName());
-			Files.delete(specPath);
-			newSpec.write(fileSystem.getPath(newSpec.getModelFileName()));
+			Path specPath = fileSystem.getPath(SpecificationWriter.getModelFileName());
+			if(Files.exists(specPath)) Files.delete(specPath);
+			// older models had a n2v.model.yaml file, delete them too
+			Path oldSpecPath = fileSystem.getPath("n2v.model.yaml");
+			if(Files.exists(oldSpecPath)) Files.delete(oldSpecPath);
+			SpecificationWriter.write(newSpec, fileSystem.getPath(SpecificationWriter.getModelFileName()));
 		}
-		return modelZooService.open(destination);
+		return modelZooService.io().open(destination);
 	}
 
 	private String getAbsolutePath(Location source) {
